refactor: share messages initiator logic
Deduplicate messages initiator detection and add safe SSE event defaults.
